refactor: split file structure into 'features' and 'common' for better code management, add MovieList component
