+++ src/App.js
@@ -1,16 +1,7 @@
-import { useEffect } from "react";
-import { useDispatch } from "react-redux";
-
-import { fetchMovies } from "./movieSlice";
+import MovieList from "./features/movieList/MovieList";
 
 function App() {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(fetchMovies());
-  }, [dispatch]);
-
-  return <p>Hello world!</p>;
+  return <MovieList />;
 }
 
 export default App;
