+++ src/common/LoadingPage/styled.js
