+++ src/features/movieList/MovieList.js
@@ -1,0 +1,16 @@
+import { useEffect } from "react";
+import { useDispatch } from "react-redux";
+
+import { fetchMovies } from "./movieSlice";
+
+function MovieList() {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchMovies());
+  }, [dispatch]);
+
+  return <h1>MovieList</h1>;
+}
+
+export default MovieList;
