+++ src/rootSaga.js
@@ -1,6 +1,6 @@
 import { all, fork } from "redux-saga/effects";
 
-import { watchFetchMoviesSaga } from "./watchFetchMoviesSaga";
+import { watchFetchMoviesSaga } from "./features/movieList/watchFetchMoviesSaga";
 
 export function* rootSaga() {
   yield all([fork(watchFetchMoviesSaga)]);
