+++ src/store.js
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "redux-saga";
 
-import moviesReducer from "./movieSlice";
+import moviesReducer from "./features/movieList/movieSlice";
 import { rootSaga } from "./rootSaga";
 
 const sagaMiddleware = createSagaMiddleware();
